requirements.txt
anyio==3.3.0
certifi==2021.5.30
click==8.0.1
dataclasses==0.6
Flask==2.0.1
Flask-Caching==1.10.1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.6
httpx==0.18.2
hyperframe==5.2.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-anilist @ git+https://github.com/0x16c3/python-anilist.git#egg=python-anilist
rfc3986==1.5.0
sniffio==1.2.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
contextvars
async-generator